refactor(operator): simplify infra construction
Remove redundant factory function, now there is one normal constructor
and one for the stub. No need to check if injected k8 client is nil
before creating the stub because the code in the stub covers that check
internally.

Give the k8 infrastructure wrapper a clear name for what its supposed to
be i.e. a k8 CLI, no need for the "wrapper" word.

Update naming everywhere. Keep only one condition in the test setup
where we need to pick if we want the stub to have a real k8 client.